+++ include/fluxgl/fluxgl.h
@@ -1,4 +1,21 @@
 #pragma once
 
+#include "core/app.h"
 #include "core/window.h"
 #include "core/error.h"
+#include "core/log.h"
+
+#include "utils/file.h"
+
+#include "input/input_manager.h"
+
+#include "graphics/vertex.h"
+#include "graphics/shader.h"
+#include "graphics/mesh.h"
+#include "graphics/texture.h"
+#include "graphics/material.h"
+#include "graphics/renderer.h"
+
+#include "components/transform.h"
+#include "components/renderable.h"
+#include "components/camera.h"