+++ include/fluxgl/utils/file.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+
+namespace fluxgl {
+    std::string readFileToString(const std::string& filePath);
+}