+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-02-18
+
+### Added
+- Minimal logging system with debug, info, warning, and error levels
+- Mesh class with vertex and index buffers
+- Shader class with vertex and fragment shaders
+- File reading system for assets (to string only for now)
+- Renderer class for meshes with shaders while still decoupling stuff
+- Texture class for loading images as textures
+- Material struct used by a renderer to draw meshes (geometry) with a given material (shader, texture, color...)
+    - Keep in mind : there is still a fair amount of coupling between shaders and renderer/materials. the API will evolve over time
+- Addded an "App" class to wrap the fluxgl functionnality (and updated examples to use this)
+- Added a "Transform" component, and a "Renderable" component to hold Transform/Mesh/Material and improve the Renderer pipeline
+- Added a "Camera" component to hold a Transform and build projection/view matrices
+- Added a "Cube" mesh and a "Sphere" mesh with a "3d" example file to show these off
+- Window glfw callbacks and input manager
+
 ## [0.1.1] - 2026-02-11
 
 ### Fixed